requirements.txt
Django
django-debug-toolbar
django-crispy-forms
psycopg2
django_stackoverflow_trace